2-1

+++ goods/src/main/java/cn/edu/xmu/oomall/goods/model/bo/OnSale.java
@@ -37,7 +37,7 @@ public class OnSale  implements  VoObject, Serializable {
 
 
     public Type getType() {
-        return Type.getTypeByCode(Integer.valueOf(type));
+        return Type.getTypeByCode(Byte.valueOf(type));
     }
 
 
@@ -48,12 +48,12 @@ public class OnSale  implements  VoObject, Serializable {
 
 
     public State getState() {
-        return State.getStatusByCode(Integer.valueOf(state));
+        return State.getStatusByCode((byte) state);
     }
 
 
     public void setState(State state) {
-        Integer code=state.getCode();
+        Byte code=state.getCode();
         Byte b=code.byteValue();
         this.state=b;
     }
@@ -70,13 +70,13 @@ public class OnSale  implements  VoObject, Serializable {
 
 
     public enum Type {
-        NOACTIVITY(0, "无活动"),
-        SECKILL(1, "秒杀"),
-        GROUPON(2, "团购"),
-        PRESALE(3, "预售");
+        NOACTIVITY((byte) 0, "无活动"),
+        SECKILL((byte) 1, "秒杀"),
+        GROUPON((byte) 2, "团购"),
+        PRESALE((byte) 3, "预售");
 
 
-        private static final Map<Integer, Type> TYPE_MAP;
+        private static final Map<Byte, Type> TYPE_MAP;
 
         static { 
             TYPE_MAP = new HashMap();
@@ -85,19 +85,19 @@ public class OnSale  implements  VoObject, Serializable {
             }
         }
 
-        private int code;
+        private Byte code;
         private String description;
 
-        Type(int code, String description) {
+        Type(Byte code, String description) {
             this.code = code;
             this.description = description;
         }
 
-        public static Type getTypeByCode(Integer code) {
+        public static Type getTypeByCode(Byte code) {
             return TYPE_MAP.get(code);
         }
 
-        public Integer getCode() {
+        public Byte getCode() {
             return code;
         }
 
@@ -109,12 +109,12 @@ public class OnSale  implements  VoObject, Serializable {
 
 
     public enum State {
-        DRAFT(0, "草稿"),
-        ONLINE(1, "上线"),
-        OFFLINE(2, "下线");
+        DRAFT((byte) 0, "草稿"),
+        ONLINE((byte) 1, "上线"),
+        OFFLINE((byte) 2, "下线");
 
 
-        private static final Map<Integer, State> STATE_MAP;
+        private static final Map<Byte, State> STATE_MAP;
 
         static { //由类加载机制，静态块初始加载对应的枚举属性到map中，而不用每次取属性时，遍历一次所有枚举值
             STATE_MAP = new HashMap();
@@ -123,19 +123,19 @@ public class OnSale  implements  VoObject, Serializable {
             }
         }
 
-        private int code;
+        private Byte code;
         private String description;
 
-        State(int code, String description) {
+        State(Byte code, String description) {
             this.code=code;
             this.description=description;
         }
 
-        public static State getStatusByCode(Integer code){
+        public static State getStatusByCode(Byte code){
             return STATE_MAP.get(code);
         }
 
-        public Integer getCode(){
+        public Byte getCode(){
             return code;
         }
